re-enable TT

+++ ai/bot/elite.go
@@ -103,7 +103,7 @@ func endGameBest(ctx context.Context, p *BotTurnPlayer, endgamePlies int) (*move
 	if err != nil {
 		return nil, err
 	}
-	log.Debug().Float32("best-endgame-val", v).Interface("seq", seq).Msg("endgame-solve-done")
+	log.Debug().Int16("best-endgame-val", v).Interface("seq", seq).Msg("endgame-solve-done")
 	return seq[0], nil
 }
